requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pydantic==2.11.3
pydantic-settings==2.8.1
pymysql==1.1.1
python-multipart==0.0.20
google-cloud-storage==2.19.0
google-cloud-vision==3.10.1
rapidfuzz==3.13.0
pillow==11.2.1
